removed some unused files and console.log lines
