+++ src/components/SpaceListAlt.jsx
@@ -20,13 +20,11 @@ const SpaceListAlt = ({ spaces, plants }) => {
             Sorry, no spaces to display!
           </p>
         ) : (
-          // add link to the NewArea component here
           <>
             <Dropdown data-bs-theme="light">
               <Dropdown.Toggle
                 id="dropdown-button-dark-example1"
                 variant="secondary"
-                // style={{ backgroundColor: "#53835C" }}
                 className="btn btn-success"
               >
                 Select a Space
@@ -36,7 +34,6 @@ const SpaceListAlt = ({ spaces, plants }) => {
                 {spaces.map((space) => {
                   return (
                     <Dropdown.Item
-                      // style={{ backgroundColor: "#A2FFB3", color: "#426949" }}\
                       className="bg-success bg-opacity-25"
                       onClick={(e) => {
                         e.preventDefault();
